Switch to double

+++ src/extended/E_Pixel.cpp
@@ -3,40 +3,40 @@
 #include "E_Pixel.h"
 
 E_Pixel::E_Pixel() {
-	m_r = 0.f;
-	m_g = 0.f;
-	m_b = 0.f;
+	m_r = 0.;
+	m_g = 0.;
+	m_b = 0.;
 }
 
-E_Pixel::E_Pixel(const float& r, const float& g, const float& b) {
+E_Pixel::E_Pixel(const double& r, const double& g, const double& b) {
 	m_r = r;
 	m_g = g;
 	m_b = b;
 }
 
 E_Pixel::E_Pixel(const olc::Pixel& pixel) {
-	m_r = static_cast<float>(pixel.r);
-	m_g = static_cast<float>(pixel.g);
-	m_b = static_cast<float>(pixel.b);
+	m_r = static_cast<double>(pixel.r);
+	m_g = static_cast<double>(pixel.g);
+	m_b = static_cast<double>(pixel.b);
 }
 
 E_Pixel::E_Pixel(const uint8_t& r, const uint8_t& g, const uint8_t& b) {
-	m_r = static_cast<float>(r);
-	m_g = static_cast<float>(g);
-	m_b = static_cast<float>(b);
+	m_r = static_cast<double>(r);
+	m_g = static_cast<double>(g);
+	m_b = static_cast<double>(b);
 }
 
 olc::Pixel E_Pixel::GetOLC() const {
-	float r = std::clamp(m_r, 0.f, 255.f);
-	float g = std::clamp(m_g, 0.f, 255.f);
-	float b = std::clamp(m_b, 0.f, 255.f);
+	double r = std::clamp(m_r, 0., 255.);
+	double g = std::clamp(m_g, 0., 255.);
+	double b = std::clamp(m_b, 0., 255.);
 	return olc::Pixel(static_cast<uint8_t>(r), static_cast<uint8_t>(g), static_cast<uint8_t>(b));
 }
 
 E_Pixel& E_Pixel::operator=(const olc::Pixel& other) {
-	m_r = static_cast<float>(other.r);
-	m_g = static_cast<float>(other.g);
-	m_b = static_cast<float>(other.b);
+	m_r = static_cast<double>(other.r);
+	m_g = static_cast<double>(other.g);
+	m_b = static_cast<double>(other.b);
 	return *this;
 }
 
@@ -50,16 +50,16 @@ E_Pixel& E_Pixel::operator=(const E_Pixel& other) {
 }
 
 E_Pixel& E_Pixel::operator/=(const E_Pixel& other) {
-	m_r = (255.f * m_r) / other.m_r;
-	m_g = (255.f * m_g) / other.m_g;
-	m_b = (255.f * m_b) / other.m_b;
+	m_r = (255. * m_r) / other.m_r;
+	m_g = (255. * m_g) / other.m_g;
+	m_b = (255. * m_b) / other.m_b;
 	return *this;
 }
 
 E_Pixel& E_Pixel::operator*=(const E_Pixel& other) {
-	m_r = (m_r * other.m_r) / 255.f;
-	m_g = (m_g * other.m_g) / 255.f;
-	m_b = (m_b * other.m_b) / 255.f;
+	m_r = (m_r * other.m_r) / 255.;
+	m_g = (m_g * other.m_g) / 255.;
+	m_b = (m_b * other.m_b) / 255.;
 	return *this;
 }
 
@@ -77,14 +77,14 @@ E_Pixel& E_Pixel::operator+=(const E_Pixel& other) {
 	return *this;
 }
 
-E_Pixel& E_Pixel::operator/=(const float& scalar) {
+E_Pixel& E_Pixel::operator/=(const double& scalar) {
 	m_r /= scalar;
 	m_g /= scalar;
 	m_b /= scalar;
 	return *this;
 }
 
-E_Pixel& E_Pixel::operator*=(const float& scalar) {
+E_Pixel& E_Pixel::operator*=(const double& scalar) {
 	m_r *= scalar;
 	m_g *= scalar;
 	m_b *= scalar;
@@ -92,19 +92,19 @@ E_Pixel& E_Pixel::operator*=(const float& scalar) {
 }
 
 void E_Pixel::Clamp() {
-	m_r = std::clamp(m_r, 0.f, 255.f);
-	m_g = std::clamp(m_g, 0.f, 255.f);
-	m_b = std::clamp(m_b, 0.f, 255.f);
+	m_r = std::clamp(m_r, 0., 255.);
+	m_g = std::clamp(m_g, 0., 255.);
+	m_b = std::clamp(m_b, 0., 255.);
 }
 
 void E_Pixel::Dither(const int& x, const int& y, const int& factor) {
-	const float f_factor = static_cast<float>(factor);
+	const double f_factor = static_cast<double>(factor);
 
-	float threshold = static_cast<float>(E_Pixel::Threshold[(x % 16) + (y % 16) * 16]) / 256.f;
-	E_Pixel octet(1.f / f_factor, 1.f / f_factor, 1.f / f_factor);
+	double threshold = static_cast<double>(E_Pixel::Threshold[(x % 16) + (y % 16) * 16]) / 256.;
+	E_Pixel octet(1. / f_factor, 1. / f_factor, 1. / f_factor);
 
-	octet *= 255.f;
-	octet *= threshold - 0.5f;
+	octet *= 255.;
+	octet *= threshold - 0.5;
 
 	this->operator+=(octet);
 	this->operator*=(f_factor);
@@ -116,84 +116,84 @@ void E_Pixel::Dither(const int& x, const int& y, const int& factor) {
 }
 
 void E_Pixel::Round() {
-	m_r = std::roundf(m_r / 255.f) * 255.f;
-	m_g = std::roundf(m_g / 255.f) * 255.f;
-	m_b = std::roundf(m_b / 255.f) * 255.f;
+	m_r = std::round(m_r / 255.) * 255.;
+	m_g = std::round(m_g / 255.) * 255.;
+	m_b = std::round(m_b / 255.) * 255.;
 }
 
 void E_Pixel::LinearToSRGB() {
-	m_r /= 255.f;
-	m_g /= 255.f;
-	m_b /= 255.f;
+	m_r /= 255.;
+	m_g /= 255.;
+	m_b /= 255.;
 
-	if (m_r <= 0.0031308f) {
-		m_r *= 12.92f;
+	if (m_r <= 0.0031308) {
+		m_r *= 12.92;
 	}
 	else {
-		m_r = ((std::powf(m_r, 1.f / 2.4f)) * 1.055f) - 0.055f;
+		m_r = ((std::pow(m_r, 1. / 2.4)) * 1.055) - 0.055;
 	}
 
-	if (m_g <= 0.0031308f) {
-		m_g *= 12.92f;
+	if (m_g <= 0.0031308) {
+		m_g *= 12.92;
 	}
 	else {
-		m_g = ((std::powf(m_g, 1.f / 2.4f)) * 1.055f) - 0.055f;
+		m_g = ((std::pow(m_g, 1. / 2.4)) * 1.055) - 0.055;
 	}
 
-	if (m_b <= 0.0031308f) {
-		m_b *= 12.92f;
+	if (m_b <= 0.0031308) {
+		m_b *= 12.92;
 	}
 	else {
-		m_b = ((std::powf(m_b, 1.f / 2.4f)) * 1.055f) - 0.055f;
+		m_b = ((std::pow(m_b, 1. / 2.4)) * 1.055) - 0.055;
 	}
 
-	m_r *= 255.f;
-	m_g *= 255.f;
-	m_b *= 255.f;
+	m_r *= 255.;
+	m_g *= 255.;
+	m_b *= 255.;
 }
 
 void E_Pixel::SRGBtoLinear() {
-	m_r /= 255.f;
-	m_g /= 255.f;
-	m_b /= 255.f;
+	m_r /= 255.;
+	m_g /= 255.;
+	m_b /= 255.;
 
-	if (m_r <= 0.04045f) {
-		m_r /= 12.92f;
+	if (m_r <= 0.04045) {
+		m_r /= 12.92;
 	}
 	else {
-		m_r = std::powf((m_r + 0.055f) / 1.055f, 2.4f);
+		m_r = std::pow((m_r + 0.055) / 1.055, 2.4);
 	}
 
-	if (m_g <= 0.04045f) {
-		m_g /= 12.92f;
+	if (m_g <= 0.04045) {
+		m_g /= 12.92;
 	}
 	else {
-		m_g = std::powf((m_g + 0.055f) / 1.055f, 2.4f);
+		m_g = std::pow((m_g + 0.055) / 1.055, 2.4);
 	}
 
-	if (m_b <= 0.04045f) {
-		m_b /= 12.92f;
+	if (m_b <= 0.04045) {
+		m_b /= 12.92;
 	}
 	else {
-		m_b = std::powf((m_b + 0.055f) / 1.055f, 2.4f);
+		m_b = std::pow((m_b + 0.055) / 1.055, 2.4);
 	}
 
-	m_r *= 255.f;
-	m_g *= 255.f;
-	m_b *= 255.f;
+	m_r *= 255.;
+	m_g *= 255.;
+	m_b *= 255.;
 }
 
-E_Pixel E_Pixel::Cubic(const E_Pixel& a, const E_Pixel& b, const E_Pixel& c, const E_Pixel& d, const float& factor) {
-	float r = Maths::Cubic(a.m_r, b.m_r, c.m_r, d.m_r, factor);
-	float g = Maths::Cubic(a.m_g, b.m_g, c.m_g, d.m_g, factor);
-	float c_b = Maths::Cubic(a.m_b, b.m_b, c.m_b, d.m_b, factor);
+E_Pixel E_Pixel::Cubic(const E_Pixel& a, const E_Pixel& b, const E_Pixel& c, const E_Pixel& d, const double& factor) {
+	double r = Maths::Cubic(a.m_r, b.m_r, c.m_r, d.m_r, factor);
+	double g = Maths::Cubic(a.m_g, b.m_g, c.m_g, d.m_g, factor);
+	double c_b = Maths::Cubic(a.m_b, b.m_b, c.m_b, d.m_b, factor);
 
 	return E_Pixel(r, g, c_b);
 }
 
-E_Pixel E_Pixel::Linear(const E_Pixel& a, const E_Pixel& b, const float& factor) {
-	float r = Maths::Map(factor, 0.f, 1.f, a.m_r, b.m_r);
-	float g = Maths::Map(factor, 0.f, 1.f, a.m_g, b.m_g);
-	float c_b = Maths::Map(factor, 0.f, 1.f, a.m_b, b.m_b);
+E_Pixel E_Pixel::Linear(const E_Pixel& a, const E_Pixel& b, const double& factor) {
+	double r = Maths::Map(factor, 0., 1., a.m_r, b.m_r);
+	double g = Maths::Map(factor, 0., 1., a.m_g, b.m_g);
+	double c_b = Maths::Map(factor, 0., 1., a.m_b, b.m_b);
 	return E_Pixel(r, g, c_b);
 }
